+++ CloudHarvestApi/blueprints/__init__.py
@@ -1,3 +1,3 @@
-from .cache.routes import cache_blueprint
+from .persistent_silo.routes import persistent_cache_blueprint
 from .home.routes import home_blueprint
 from .reports.routes import report_blueprint