+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.1.12
+- Update to conform to CloudHarvestCoreTasks 0.4.0
+- Added the `redis` service to `docker-compose.yaml`
+- Moved several persistent cache operations to the ephemeral cache
+- Moved `BaseCacheTask` and `CacheAggregateTask` to `CloudHarvestCoreTasks` as refactors inheriting `BaseDataTask`
+
 # 0.1.11
 - Updated `docker-compose` so that `mongo` manages its own data volume
 - Fixed invalid `task_chain_from_dict()` arguments in `reports/routes.py`
@@ -18,7 +24,7 @@
 - Removed some unused lines from `docker-compose.yaml`
 
 # 0.1.6
-- Changed the output of the [`report_run()`](CloudHarvestApi/api/blueprints/reports/routes.py) method from `dict` to `List[dict]`
+- Changed the output of the [`report_run()`](CloudHarvestApi/blueprints/reports/routes.py) method from `dict` to `List[dict]`
 - Updated `BaseCacheTask` to accept the `title` parameter in the constructor
 
 # 0.1.5